+++ src/main/java/duke/commands/AddTaskCommand.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.tasks.Task;
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 public class AddTaskCommand extends Command {
     Task task;
     public AddTaskCommand(Task task) {
@@ -7,5 +14,5 @@ public class AddTaskCommand extends Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         tasks.addNewTask(task);
         ui.showAddTask(tasks, task);
-    };
+    }
 }
