+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+import duke.DukeException;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
