+++ src/main/java/duke/commands/DoneCommand.java
@@ -1,3 +1,11 @@
+package duke.commands;
+
+import duke.DukeException;
+import duke.tasks.Task;
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 public class DoneCommand extends Command {
     private final int taskIndex;
     public DoneCommand(int taskIndex) {
