+++ src/main/java/duke/tasks/Deadline.java
@@ -1,17 +1,20 @@
+package duke.tasks;
+
+import duke.tasks.Task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task {
-
+public class Deadline extends Task {
     private static DateTimeFormatter FORMATTER = DateTimeFormatter.ofPattern("d-M-uuuu H:mm");
     private static DateTimeFormatter DISPLAY_FORMATTER = DateTimeFormatter.ofPattern("d MMM uuuu hh:mm a");
     protected LocalDateTime date;
 
-    public Event(String description, String at, boolean isDone) {
+    public Deadline(String description, String by, boolean isDone) {
         super(description, isDone);
-        this.date = parseDateTime(at);
+        this.date = parseDateTime(by);
     }
-    
+
     private LocalDateTime parseDateTime(String at) {
         // TODO: need to catch exception
         return LocalDateTime.parse(at, FORMATTER);
@@ -19,11 +22,11 @@ public class Event extends Task {
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (at: " + date.format(DISPLAY_FORMATTER) + ")";
+        return "[D]" + super.toString() + " (by: " + date.format(DISPLAY_FORMATTER) + ")";
     }
 
     @Override
     public String toFileString() {
-        return "E" + super.toFileString() + " | " + date.format(FORMATTER);
+        return "D" + super.toFileString() + " | " + date.format(FORMATTER);
     }
 }
